+++ client/src/utils/queries.js
@@ -57,11 +57,6 @@ export const GET_ALL_ITEMS = gql`
       serial
       categories {
         _id
-        category
-      }
-      tradeFor {
-        _id
-        category
       }
       expire
       dateListed
@@ -109,7 +104,7 @@ export const GET_ALL_CATEGORIES = gql`
   query {
     getAllCategories {
       _id
-      category
+      name
     }
   }
 `;
